Relocate "Recordings" directory inside "Documents"
Instead of locate the "Recordings" directory inside the $HOME dir,
we have been able to move this directory to the XDG directory
"$ HOME / Documents" because it is considered that this type of
file is somehow a document.

On the other hand, if there was a previous installation and the
directory "Recordings" was within "$HOME" it will preserve the
backward compatibility but also it will be saved as a setting in
case you want to change the default directory.

https://bugzilla.gnome.org/show_bug.cgi?id=748092